pyproject.toml
[project]
name = "qna-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.24.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.59",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
